% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{as_connection_contract}
\alias{as_connection_contract}
\title{Coerces object into an RStudio connection contract}
\usage{
as_connection_contract(x)
}
\arguments{
\item{x}{Object that will be coerced}
}
\description{
Coerces object into an RStudio connection contract
}
