% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_close.R
\name{connection_close}
\alias{connection_close}
\title{Close a connection}
\usage{
connection_close(con, leave_open = FALSE)
}
\arguments{
\item{con}{Connection variable}

\item{leave_open}{Should the connection be left open. Defaults to FALSE.}
}
\description{
Close a connection
}
\examples{
library(DBI)
con <- connection_open(RSQLite::SQLite(), path = ":dbname:")
connection_close(con)
con
}
