% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{connection_list}
\alias{connection_list}
\title{Creates an RStudio IDE Contract object}
\usage{
connection_list(spec = base_spec())
}
\arguments{
\item{spec}{A list object that containing the structure of the Contract object}
}
\description{
Creates an RStudio IDE Contract object
}
\examples{

connection_list()

}
