% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_open.R
\name{connection_open}
\alias{connection_open}
\title{Opens a connection}
\usage{
connection_open(...)
}
\arguments{
\item{...}{Passes arguments to wrapped connection function}
}
\description{
Opens a connection
}
\examples{
library(DBI)
con <- connection_open(RSQLite::SQLite(), path = ":dbname:")
con
connection_close(con)
}
