% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pin-read.R
\name{connection_pin_read}
\alias{connection_pin_read}
\title{Retrieves a database connection or query from a board}
\usage{
connection_pin_read(board, name, version = NULL)
}
\arguments{
\item{board}{A `pins` board object}

\item{name}{The name of the pin}

\item{version}{The version of the pin to get (optional)}
}
\description{
Retrieves a database connection or query from a board
}
