% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pin-write.R
\name{connection_pin_write}
\alias{connection_pin_write}
\title{Writes a database connection or query to a board}
\usage{
connection_pin_write(board, x, ...)
}
\arguments{
\item{board}{A `pins` board object}

\item{x}{A `connections` table or database connection}

\item{...}{Additional arguments to pass to `pins::pin_write()`}
}
\description{
Writes a database connection or query to a board
}
