% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_update.R
\name{connection_update}
\alias{connection_update}
\title{Refreshes a connection}
\usage{
connection_update(con, hint = "")
}
\arguments{
\item{con}{Connection variable}

\item{hint}{Optional argument passed to the Contract}
}
\value{
Returns a NULL object. If using the RStudio IDE, it will attempt to refresh the
connection identified by attributes of the con object
}
\description{
Refreshes a connection
}
\examples{
library(DBI)
con <- connection_open(RSQLite::SQLite(), path = ":dbname:")
connection_update(con)
connection_close(con)

}
