% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_methods.R
\docType{methods}
\name{dbWriteTable,connections_class,ANY-method}
\alias{dbWriteTable,connections_class,ANY-method}
\title{Copy data frames to database tables}
\usage{
\S4method{dbWriteTable}{connections_class,ANY}(conn, name, value, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{name}{A character string specifying the unquoted DBMS table name,
or the result of a call to \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{...}{Other parameters passed on to methods.}
}
\description{
Copy data frames to database tables
}
