% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pin-write.R
\name{write_pin_conn}
\alias{write_pin_conn}
\title{Writes pin (internal)}
\usage{
write_pin_conn(x, board, ...)
}
\arguments{
\item{x}{A `connections` table or database connection}

\item{board}{A `pins` board object}

\item{...}{Additional arguments to pass to `pins::pin_write()`}
}
\description{
Writes pin (internal)
}
\keyword{internal}
