% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-code.R
\name{connection_code}
\alias{connection_code}
\title{Displays the code that will be used to recreate the connection}
\usage{
connection_code(con)
}
\arguments{
\item{con}{A `connConnection` object}
}
\value{
It prints out the lines of code that this package will use
to reconnect to the database.
}
\description{
Displays the code that will be used to recreate the connection
}
