% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{connection_contract}
\alias{connection_contract}
\title{Creates an RStudio IDE Contract object}
\usage{
connection_contract(spec = base_spec())
}
\arguments{
\item{spec}{A list object that containing the structure of the Contract object}
}
\description{
Creates an RStudio IDE Contract object
}
\examples{

connection_contract()
}
