% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-open.R
\name{connection_open}
\alias{connection_open}
\title{Opens a connection}
\usage{
connection_open(..., open_pane = TRUE)
}
\arguments{
\item{...}{Passes arguments to wrapped connection function}

\item{open_pane}{Signals for the RStudio Connections pane to open.
Defaults to TRUE.}
}
\value{
Returns a NULL object.  If using the RStudio IDE, it will attempt to open the
connection
}
\description{
Opens a connection
}
\examples{
library(DBI)
con <- connection_open(RSQLite::SQLite(), path = ":dbname:")
con
connection_close(con)
}
