% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-view.R
\name{connection_view}
\alias{connection_view}
\title{Populates the RStudio Connection viewer}
\usage{
connection_view(
  con,
  connection_code = "",
  host = "",
  name = "",
  connection_id = ""
)
}
\arguments{
\item{con}{Connection variable}

\item{connection_code}{Text of code to connect to the same source}

\item{host}{Name of Host of the connection}

\item{name}{Connection name}

\item{connection_id}{Unique ID of the connection for the current session}
}
\description{
Populates the RStudio Connection viewer
}
\examples{
library(DBI)
con <- connection_open(RSQLite::SQLite(), path = ":dbname:")
connection_view(con)
connection_close(con)
}
